refactor: Organize config files into dedicated package
Moved config-related files into src/keecas/config/ folder structure:
- config.py → config/manager.py
- config_migration.py → config/migration.py
- config_schema.py → config/schema.py
- Created config/__init__.py with proper exports

Updated all imports across:
- cli.py: Updated import paths for config commands
- display.py: Updated manager and EnvironmentDefinition imports
- pint_sympy.py: Updated get_config_manager import
- tests/test_config_migration.py: Updated test imports

All tests passing (24/24). CLI commands verified working.
